allowed host

+++ orders/views.py
@@ -11,16 +11,23 @@ from django.views.decorators.csrf import csrf_exempt
 from django.http import HttpResponseBadRequest
 from django.core.mail import EmailMessage
 from offer.models import Coupon,RedeemedCoupon
+from forex_python.converter import CurrencyRates
 from django.template.loader import render_to_string
 from django.http import JsonResponse
 razorpay_client=razorpay.Client(
     auth=(settings.RAZOR_KEY_ID,settings.RAZOR_KEY_SECRET))
 
+def dollar_rate():
+    c = CurrencyRates()
+    rate = c.get_rate('USD', 'INR')
+    return rate
+
+
 
 # Create your views here.
 def orders(request,total=0,quantity=0):
     current_user= request.user
-
+    item=None
     # if cart count is less than  or equal to zero
     cart_items = CartItem.objects.filter(user=current_user)
     cart_count = cart_items.count()
@@ -33,35 +40,47 @@ def orders(request,total=0,quantity=0):
             coupon_redeem.user=current_user
             coupon_redeem.coupon=coupon
             coupon_redeem.save()
-    if cart_count <= 0:
-        return redirect('store')
+    # if cart_count <= 0:
+    #     return redirect('store')
 
     grand_total =0
     tax=0
     total_savings=0
     offer_savings=0
-    item=None
-    for cart_item in cart_items:
-        total += (cart_item.product.get_price() * cart_item.quantity)
-        offer_savings=total-(cart_item.product.price*cart_item.quantity)
-        quantity += cart_item.quantity
-    tax=(2*total)/100
-    grand_total=total+tax
-
-    if request.session.has_key('couponid'):
-        print('inside Coupen')
-        # couponid=request.session['coupon_id']
-        # request.session['couponid']=couponid
-        # del request.session['coupon_id']
-        coupen_discount= request.session['coupon_discount']
-        coupen_discount_price = total*(coupen_discount)/100
-        grand_total=grand_total-coupen_discount_price
-        total_savings=grand_total-(coupen_discount_price-offer_savings)
-        print(grand_total)
+    if 'direct_buy' in request.session:
+        product=request.session['direct_buy']
+        item=Product.objects.get(id=product)
+        tax=(2*item.get_price())/100
+        total=item.get_price()
+        quantity=1
+        offer_savings=total-item.price
+        grand_total=total+tax
+        if request.session.has_key('couponid'):
+            print('inside Coupen')
+            coupen_discount= request.session['coupon_discount']
+            coupen_discount_price = total*(coupen_discount)/100
+            grand_total=grand_total-coupen_discount_price
+            total_savings=grand_total-offer_savings
+            print(grand_total)
+        else:
+            coupen_discount_price=0
     else:
-        coupen_discount_price=0
-
-    total_savings= grand_total-(coupen_discount_price-offer_savings)
+        for cart_item in cart_items:
+            total += (cart_item.product.get_price() * cart_item.quantity)
+            offer_savings=total-(cart_item.product.price*cart_item.quantity)
+            quantity += cart_item.quantity
+            total_savings=grand_total-offer_savings
+        tax=(2*total)/100
+        grand_total=total+tax
+        if request.session.has_key('couponid'):
+                print('inside Coupen')
+                coupen_discount= request.session['coupon_discount']
+                coupen_discount_price = total*(coupen_discount)/100
+                grand_total=grand_total-coupen_discount_price
+                total_savings= grand_total-(coupen_discount_price-offer_savings)
+        else:
+            coupen_discount_price=0
+    paypal_amount = grand_total/dollar_rate()
     if request.method =='POST':
         print("inside if")
         form =OrderForm(request.POST)
@@ -94,25 +113,30 @@ def orders(request,total=0,quantity=0):
             data.save()
             request.session['order_number']=order_number
             request.session['grand_total']=grand_total
-            order_amount = int(grand_total*100)
+            order_amount = (grand_total*100)
             print(order_amount)
             order_currency = 'INR'
-            razorpay_order = razorpay_client.order.create(dict(amount=order_amount,currency=order_currency,payment_capture='0'))
+            razorpay_order = razorpay_client.order.create(dict(amount=int(order_amount),currency=order_currency,payment_capture='0'))
             payment_order_id = razorpay_order['id']
-            callback_url = 'paymenthandler/'
+            # callback_url = 'paymenthandler/'
 
             order= Ordern.objects.get(user=current_user,is_ordered=False,order_number=order_number)
             context={
                 'order':order,
                 'cart_items':cart_items,
-                'total':total,
-                'tax':tax,
+                'total':round(total),
+                'tax':round(tax,2),
+                'item':item,
+                'quantity':quantity,
                 'grand_total':grand_total,
                 'payment_order_id': payment_order_id,
                 'razorpay_amount':order_amount,
                 'currency':order_currency,
-                'callback_url' : callback_url,
+                'razorpay_merchant_key':settings.RAZOR_KEY_ID,
+                # 'callback_url' : callback_url,
                 'total_savings':total_savings,
+                'paypal_amount':round(paypal_amount,2),
+                'currency':order_currency,
                 'coupen_discount_price':coupen_discount_price,
             }
             return render(request,'store/payments.html',context)
@@ -155,7 +179,7 @@ def paymenthandler(request):
                     # if there is an error while capturing payment.
                     return render(request, 'paymentfail.html')
             else:
- 
+                
                 # if signature verification fails.
                 return render(request, 'paymentfail.html')
         except:
@@ -178,7 +202,7 @@ def payments(request):
     payment =Payment(
         user = request.user,
         payment_id = body['transID'],
-        payment_method = body['payment_method'],
+        payment_method = 'paypal',
         amount_paid = order.order_total,
         status = body['status'],
     )
@@ -187,34 +211,56 @@ def payments(request):
     order.is_ordered=True
     order.save()
 
-    #Move cart items to Order Product table
-    cart_items=CartItem.objects.filter(user=request.user)
-
-    for item in cart_items:
-        orderproduct=OrderProduct()
-        orderproduct.order_id =order.id
-        orderproduct.payment =payment
+    if 'direct_buy' in request.session:
+        product_id=request.session['direct_buy']
+        direct_item=Product.objects.get(id=product_id)
+        item=Product.objects.get(id=product_id)
+        orderproduct = OrderProduct()
+        orderproduct.order_id = order.id
         orderproduct.user_id = request.user.id
-        orderproduct.product_id = item.product_id
-        orderproduct.quantity = item.quantity
-        orderproduct.product_price=item.product.price
-        orderproduct.ordered =True
+        orderproduct.product_id = product_id
+        orderproduct.quantity = 1
+        orderproduct.product_price = direct_item.price
+        orderproduct.ordered = True
         orderproduct.save()
 
-        cart_item = CartItem.objects.get(id=item.id)
-        product_variation = cart_item.variations.all()
-        orderproduct = OrderProduct.objects.get(id=orderproduct.id)
-        orderproduct.variations.set(product_variation)
-        orderproduct.save()
-
-        #Reduce the quantity of the sold products
-
-        product=Product.objects.get(id=item.product_id)
-        product.stock -= item.quantity
+        product = Product.objects.get(id=product_id)
+        product.stock -= 1
         product.save()
 
+    else:
+
+    #Move cart items to Order Product table
+        cart_items=CartItem.objects.filter(user=request.user)
+
+        for item in cart_items:
+            orderproduct=OrderProduct()
+            orderproduct.order_id =order.id
+            orderproduct.payment =payment
+            orderproduct.user_id = request.user.id
+            orderproduct.product_id = item.product_id
+            orderproduct.quantity = item.quantity
+            orderproduct.product_price=item.product.price
+            orderproduct.ordered =True
+            orderproduct.save()
+
+            cart_item = CartItem.objects.get(id=item.id)
+            product_variation = cart_item.variations.all()
+            orderproduct = OrderProduct.objects.get(id=orderproduct.id)
+            orderproduct.variations.set(product_variation)
+            orderproduct.save()
+
+            #Reduce the quantity of the sold products
+
+            product=Product.objects.get(id=item.product_id)
+            product.stock -= item.quantity
+            product.save()
+
     # clear cart
-    CartItem.objects.filter(user=request.user).delete()
+    if 'direct_buy' in request.session:
+        del request.session['direct_buy']
+    else:
+        CartItem.objects.filter(user=request.user).delete()
 
     # send order recieved email to  customer
     mail_subject = 'Thank you for your order'
@@ -241,22 +287,21 @@ def order_complete(request):
    transID = request.GET.get('payment_id')
 
    try:
-        order = Ordern.objects.get(order_number=order_number, is_ordered=False)
-        ordered_products = OrderProduct.objects.get(order_id=order.id)
+        order = Ordern.objects.get(is_ordered=False,order_number=order_number)
+        ordered_products = OrderProduct.objects.filter(order_id=order.id)
         order.is_ordered=True
         order.save()
         subtotal = 0
         for i in ordered_products:
             subtotal += i.product_price * i.quantity
 
-        payment = Payment.objects.get(payment_id=transID)
+        # payment = Payment.objects.get(payment_id=transID)
 
         context = {
             'order': order,
             'ordered_products': ordered_products,
             'order_number': order.order_number,
-            'transID': payment.payment_id,
-            'payment': payment,
+           
             'subtotal': subtotal,
         }
         return render(request, 'order_complete.html', context)
@@ -305,7 +350,6 @@ def order_complete(request):
                         return redirect('home')
                 
        
-
 def user_orders(request,total=0,quantity=0):
     # tax=0
     # grand_total=0
@@ -367,35 +411,56 @@ def cash_on_delivery(request,total=0,quantity=0):
     # move cartitems to orderproduct table
     order_number=request.session['order_number']
     order= Ordern.objects.get(user=request.user,is_ordered=False,order_number=order_number)
-    cart_items=CartItem.objects.filter(user=request.user)
 
-
-    for item in cart_items:
-        orderproduct=OrderProduct()
-        orderproduct.order_id=order.id
-        orderproduct.user_id=request.user.id
-        orderproduct.product_id=item.product_id
-        orderproduct.quantity=item.quantity
-        orderproduct.product_price=item.product.price
-        orderproduct.ordered=True
+    if 'direct_buy' in request.session:
+        product_id=request.session['direct_buy']
+        direct_buy_item=Product.objects.get(id=product_id)
+        item=Product.objects.get(id=product_id)
+        orderproduct = OrderProduct()
+        orderproduct.order_id = order.id
+        orderproduct.user_id = request.user.id
+        orderproduct.product_id = product_id
+        orderproduct.quantity = 1
+        orderproduct.product_price = direct_buy_item.price
+        orderproduct.ordered = True
         orderproduct.save()
 
-
-        cart_item=CartItem.objects.get(id=item.id)
-        product_variation = cart_item.variations.all()
-        orderproduct = OrderProduct.objects.get(id=orderproduct.id)
-        orderproduct.variations.set(product_variation)
-        orderproduct.save()
-
-        #reduce the quantity of the product 
-        product = Product.objects.get(id=item.product_id)
-        product.stock -= item.quantity
+        product=Product.objects.get(id=product_id)
+        product.stock -= 1
         product.save()
+
+    else:
+
+        cart_items=CartItem.objects.filter(user=request.user)
+        for item in cart_items:
+            orderproduct=OrderProduct()
+            orderproduct.order_id=order.id
+            orderproduct.user_id=request.user.id
+            orderproduct.product_id=item.product_id
+            orderproduct.quantity=item.quantity
+            orderproduct.product_price=item.product.price
+            orderproduct.ordered=True
+            orderproduct.save()
+
+
+            cart_item=CartItem.objects.get(id=item.id)
+            product_variation = cart_item.variations.all()
+            orderproduct = OrderProduct.objects.get(id=orderproduct.id)
+            orderproduct.variations.set(product_variation)
+            orderproduct.save()
+
+            #reduce the quantity of the product 
+            product = Product.objects.get(id=item.product_id)
+            product.stock -= item.quantity
+            product.save()
     # order.is_ordered=True
     # order.save()
 
     #clear cart
-    CartItem.objects.filter(user=request.user).delete()
+    if 'direct_buy' in request.session:
+        del request.session['direct_buy']
+    else:
+        CartItem.objects.filter(user=request.user).delete()
     return redirect('order_complete')
 
 def payment_failed(request):
@@ -435,32 +500,53 @@ def razorpay_payment_verification(request):
         order.is_ordered = True
         order.save()
 
-    # Move the cart items to Order Product table
-    cart_items = CartItem.objects.filter(user=request.user)
-
-    for item in cart_items:
+    if 'direct_buy' in request.session:
+        product_id=request.session['direct_buy']
+        direct_buy_item=Product.objects.get(id=product_id)
+        item=Product.objects.get(id=product_id)
         orderproduct = OrderProduct()
         orderproduct.order_id = order.id
-        orderproduct.payment = payment
         orderproduct.user_id = request.user.id
-        orderproduct.product_id = item.product_id
-        orderproduct.quantity = item.quantity
-        orderproduct.product_price = item.product.price
+        orderproduct.product_id = product_id
+        orderproduct.quantity = 1
+        orderproduct.product_price = direct_buy_item.price
         orderproduct.ordered = True
         orderproduct.save()
 
-        cart_item = CartItem.objects.get(id=item.id)
-        product_variation = cart_item.variations.all()
-        orderproduct = OrderProduct.objects.get(id=orderproduct.id)
-        orderproduct.variations.set(product_variation)
-        orderproduct.save()
-
-
-    #     # Reduce the quantity of the sold products
-        product = Product.objects.get(id=item.product_id)
-        product.stock -= item.quantity
+        product=Product.objects.get(id=product_id)
+        product.stock -= 1
         product.save()
 
+    else:
+
+    # Move the cart items to Order Product table
+        cart_items = CartItem.objects.filter(user=request.user)
+        for item in cart_items:
+            orderproduct = OrderProduct()
+            orderproduct.order_id = order.id
+            orderproduct.payment = payment
+            orderproduct.user_id = request.user.id
+            orderproduct.product_id = item.product_id
+            orderproduct.quantity = item.quantity
+            orderproduct.product_price = item.product.price
+            orderproduct.ordered = True
+            orderproduct.save()
+
+            cart_item = CartItem.objects.get(id=item.id)
+            product_variation = cart_item.variations.all()
+            orderproduct = OrderProduct.objects.get(id=orderproduct.id)
+            orderproduct.variations.set(product_variation)
+            orderproduct.save()
+
+
+        #     # Reduce the quantity of the sold products
+            product = Product.objects.get(id=item.product_id)
+            product.stock -= item.quantity
+            product.save()
+
     # # Clear cart
-    CartItem.objects.filter(user=request.user).delete()
+    if 'direct_buy' in request.session:
+        del request.session['direct_buy']
+    else:
+        CartItem.objects.filter(user=request.user).delete()
     return JsonResponse({'message': 'success'})
